SmartRAG/requirements.txt
streamlit
sentence-transformers
faiss-cpu
rank-bm25
numpy
pymupdf
huggingface_hub
transformers
